Upgrade gradle version to 8.0

+++ sample/src/main/java/com/stfalcon/chatkit/sample/features/demo/DemoMessagesActivity.java
@@ -62,14 +62,11 @@ public abstract class DemoMessagesActivity extends AppCompatActivity
 
     @Override
     public boolean onOptionsItemSelected(MenuItem item) {
-        switch (item.getItemId()) {
-            case R.id.action_delete:
-                messagesAdapter.deleteSelectedMessages();
-                break;
-            case R.id.action_copy:
-                messagesAdapter.copySelectedMessagesText(this, getMessageStringFormatter(), true);
-                AppUtils.showToast(this, R.string.copied_message, true);
-                break;
+        if (item.getItemId() == R.id.action_delete) {
+            messagesAdapter.deleteSelectedMessages();
+        } else if (item.getItemId() == R.id.action_copy) {
+            messagesAdapter.copySelectedMessagesText(this, getMessageStringFormatter(), true);
+            AppUtils.showToast(this, R.string.copied_message, true);
         }
         return true;
     }
